Ora-mar-03

+++ 11 - OOP/Feladat02/motor.py
@@ -8,6 +8,6 @@ class motor():
         self.henger :int = 0
         self.gyartasiev :int = 0
         self.nyomatek :int = 0
-        self.fogyasztas :int = 0
+        self.fogyasztas :float = 0
     def __str__(self) -> str:
         return f"\nA motorkerépkár adatai:\n\tTípusa: {self.tipus}\n\tGyártja: {self.gyarto}\n\tModell: {self.modell}\n\tLóereje: {self.loero} HP\n\tHenegerek: {self.henger}\n\tGyártási év: {self.gyartasiev}\n\tNyomatéka: {self.nyomatek}\n\tFogyasztása: {self.fogyasztas} l/100km\n"
